Whitespace corrections, smfitness reporting, WT missing reporting
I had put in a number of tabs, these where changed to spaces(4)
The number of strains with missing fitnesses and NaN fitnesses in
the fitness file is now reported to the screen and the log.
If there are no WT screens (from which to calculate array variance)
the script prints an error, saves the state, and exits gracefully.