minor bugfix, update to chrom_coord file (missing gene) live testing of new linkage
bugfix was a misspelling of isempty
old coord file was missing an entry for YLR312C-B which is no longer a gene
it has since been merged with YLR313C. I made a line for it in the coord file
and gave it the same coords as YLR313C. The coord file now bears todays date.
new linkage is run but the results are not used (just saved for comparison) times
taken in linkage script are also recorded.