Linkage window problem
The automatic windows are not wide enough. Anastasia says this is one of the
reasons she used to go through the process more than twice. I've changed the
scoring scrip to do the lookup of linkage colonies, even if they're not going
to be set to NaN. This way, they can at least be included in the list of colonies
to ignore when running corrections (eg. row/col correction)

Also, I've added the new linkage tool to the repo (_new) and the script will
be using it intead of the old one. The old one is still being run as I still
need to do some analysis on the difference. I'm waiting for a properly scored
set to do it on.

Also, I've added a section which collapses all _collab (ie. MRECK!) screens
into a single batch. A warning is printed to the log. This will have to be
removed later.