linkage mask change
The old process had us completely skip the linkage step and report
zero colonies in linkage. This leaves an extremely strong signature
in the data and ends up causing tons of positive interactions on the
edge of the linkage window. Then when we scanned the linkage window
for negative interactions, the resulting window would not be large
enough.

Now we process linkage as normal (with the new script) so that the record
of linkage cols can get stored in "ignore_cols" which is passed to the
correction filters, and prevent the green regions from getting pushed up
(for example by plate normalization). They are then masked to NaN so as
not to participate in the batch correction. After batch, they are replaced
with their old values which came out of the jackknife step.

Also, the old linkage script is gone for good.