Big Coloony Filter and output_interaction_data
the output function now automatically prints a list of orfs to
aid the user when they try to load the huge file

There are several versions of the big colony filter in this version
of compute_sgascore.m. Things are not yet settled as to the parameters
and the procedure, but I haven't commited in a while. ;)