Header cleanup / default param vals / Wild_Types
Mostly this change fixes the wild-type names
which have changed to URA3control_sn4757xxx

It is now an input parameter with fg array default.
I've also put all of the settable flag checks and
defaults near the top and removed some of the front matter.
The linkage script takes the WT id in as a parameter.

The multi class stuff is still in flux. This is the NOORPHAN
version which simply ignores all batches with less than 3
queries and doesn't allow them to contribute to the
covariance matrix calculation.